remove _handler_exceptions file
