+++ services/web/server/src/simcore_service_webserver/catalog/_exceptions.py
@@ -1,11 +1,36 @@
 """Defines the different exceptions that may arise in the catalog subpackage"""
 
+import functools
+
+from aiohttp import web
+from servicelib.aiohttp.typing_extension import Handler
 from servicelib.rabbitmq.rpc_interfaces.catalog.errors import (
     CatalogForbiddenError,
     CatalogItemNotFoundError,
 )
 
 from ..errors import WebServerBaseError
+from ..resource_usage.errors import DefaultPricingPlanNotFoundError
+
+
+def reraise_catalog_exceptions_as_http_errors(handler: Handler):
+    @functools.wraps(handler)
+    async def _wrapper(request: web.Request) -> web.StreamResponse:
+        try:
+
+            return await handler(request)
+
+        except (
+            CatalogItemNotFoundError,
+            DefaultPricingPlanNotFoundError,
+            DefaultPricingUnitForServiceNotFoundError,
+        ) as exc:
+            raise web.HTTPNotFound(reason=f"{exc}") from exc
+
+        except CatalogForbiddenError as exc:
+            raise web.HTTPForbidden(reason=f"{exc}") from exc
+
+    return _wrapper
 
 
 class BaseCatalogError(WebServerBaseError):
